修复币安提交参数错误的问题

+++ exchanges/binance/binance_types.go
@@ -84,7 +84,7 @@ type RecentTrade struct {
 	ID           int64   `json:"id"`
 	Price        float64 `json:"price,string"`
 	Quantity     float64 `json:"qty,string"`
-	Time         int64   `json:"time"`
+	Time         float64 `json:"time"`
 	IsBuyerMaker bool    `json:"isBuyerMaker"`
 	IsBestMatch  bool    `json:"isBestMatch"`
 }
